Add gnome packages and start steam
Adding the gnome packages for a very basic gnome install to get rid of
the ubuntu branding.  In my ubuntu setup, I run with a minimal install
that contains a terminal and a web browser this adds the standard gnome
utilities, fonts, and icons back in.

Also started on the automated steam client install.